+++ gatsby-node.js
@@ -17,3 +17,8 @@ exports.onCreateWebpackConfig = ({ stage, actions }) => {
     },
   });
 };
+
+exports.onCreatePage = async ({ page, actions }) => {
+  const { createPage } = actions;
+  createPage(page);
+};
